+++ src/MonadIO/Process.hs
@@ -1,5 +1,5 @@
 module MonadIO.Process
- ( doProc, system, systemx, systemN, systemS, throwSig, throwSig' )
+ ( doProc, procWait, system, systemx, systemN, systemS, throwSig, throwSig' )
 where
 
 import Base1T  hiding  ( (∉) )
@@ -12,13 +12,18 @@ import ContainersPlus.Member  ( HasMember( (∉) ) )
 
 import FPath.Error.FPathError  ( AsFPathError )
 
+-- lens --------------------------------
+
+import Control.Lens.Tuple   ( _1 )
+
 -- monaderror-io -----------------------
 
-import MonadError  ( fromRight )
+import MonadError     ( fromRight )
+import MonadError.IO  ( ioThrow )
 
 -- process -----------------------------
 
-import System.Process  ( ProcessHandle, waitForProcess )
+import System.Process  ( ProcessHandle, getPid, waitForProcess )
 
 ------------------------------------------------------------
 --                     local imports                      --
@@ -29,11 +34,14 @@ import MonadIO.Error.ProcExitError    ( AsProcExitError, asProcExitError )
 import MonadIO.NamedHandle            ( stdin )
 import MonadIO.OpenFile               ( devnull )
 import MonadIO.Process.CmdSpec        ( CmdSpec, expExit )
-import MonadIO.Process.ExitStatus     ( ExitStatus( ExitSig, ExitVal ) )
+import MonadIO.Process.ExitInfo       ( ExitInfo, exitInfo )
+import MonadIO.Process.ExitStatus     ( ExitStatus( ExitSig, ExitVal )
+                                      , exitVal, exitWasSignalled )
 import MonadIO.Process.MakeProc       ( MakeProc, makeProc )
 import MonadIO.Process.MkInputStream  ( MkInputStream )
 import MonadIO.Process.Signal         ( Signal( Signal ) )
 import MonadIO.Process.OutputHandles  ( OutputHandles( slurp ) )
+import MonadIO.Process.Pid            ( Pid( Pid ), pid )
 import MonadIO.Process.ToMaybeTexts   ( ToMaybeTexts( toMaybeTexts ) )
 
 --------------------------------------------------------------------------------
@@ -48,13 +56,27 @@ exitCode (ExitFailure i) | i > 0     = ExitVal $ fromIntegral i
 {- | Take an opened process and its output handles, wait for the process, slurp
      the handles, and return the exit val and any output (as Text). -}
 
-procWait ∷ ∀ ζ ω μ . (MonadIO μ, OutputHandles ζ ω) ⇒
+procWait' ∷ ∀ ζ ω μ . (MonadIO μ, OutputHandles ζ ω) ⇒
            μ (ProcessHandle, ζ) → μ (ExitStatus, ω)
-procWait prox = do
+procWait' prox = do
   (handle, hs) ← prox
   ex ← liftIO $ waitForProcess handle
   texts ← slurp hs
   return (exitCode ex, texts)
+
+----------------------------------------
+
+{- | Take a process handle & stdin specification; wait for termination, return
+     exit status and whichever of stderr/stdout were implicitly requested by the
+     return type.
+-}
+procWait ∷ ∀ ζ ω μ . (MonadIO μ, OutputHandles ζ ω) ⇒
+           CmdSpec → μ (ProcessHandle, ζ) → μ (ExitInfo, ω)
+procWait cspec x =
+  let go r = do
+        pid_ ← getPid' (r ⊣ _1)
+        procWait' (return r) ≫ \ (t,w) → return (exitInfo t cspec pid_, w)
+   in join $ go ⊳ x
 
 ----------------------------------------
 
@@ -66,13 +88,19 @@ systemx ∷ ∀ ε ζ ω σ μ .
           AsCreateProcError ε, AsFPathError ε, AsIOError ε, MonadError ε μ) ⇒
          σ       -- ^ stdin specification
        → CmdSpec -- ^ cmd + args
-       → μ (ExitStatus, ω)
+       → μ (ExitInfo, ω)
 
-systemx inh cspec = do
-  x ← ѥ $ makeProc inh cspec
-  ѥ x ≫ \case
-    𝕷 e → join ∘ return $ throwError e
-    𝕽 r → procWait (return r)
+systemx inh cspec =
+ ѥ (makeProc inh cspec) ≫ procWait cspec
+
+{- | Get pid where we're sure the pid should be available; throws errors into
+     IO.
+ -}
+getPid' ∷ MonadIO μ ⇒ ProcessHandle → μ Pid
+getPid' h =
+  liftIO $ getPid h ≫ \ case
+    𝕹   → ioThrow ("failed to getPid from handle; already closed" ∷ 𝕋)
+    𝕵 p → return $ Pid p
 
 -- $ system defCPOpts (""∷ Text) (CmdSpec (CmdExe [absfile|/usr/bin/env|])
 --          (CmdArgs []))
@@ -94,14 +122,15 @@ system ∷ ∀ ε ζ ω σ μ .
           MonadError ε μ, HasCallStack) ⇒
          σ       -- ^ stdin specification
        → CmdSpec -- ^ cmd + args
-       → μ (ExitStatus, ω)
+       → μ (ExitInfo, ω)
 
 system inh cspec = do
-  (exit,w) ← systemx inh cspec
+  (einfo,w) ← systemx inh cspec
 
-  if exit ∉ (cspec ⊣ expExit)
-  then throwError $ asProcExitError cspec exit (toMaybeTexts w)
-  else return (exit,w)
+  if (einfo ⊣ exitVal) ∉ (cspec ⊣ expExit)
+  then let x = einfo ⊣ exitVal
+        in throwError $ asProcExitError cspec (einfo ⊣ pid) x (toMaybeTexts w)
+  else return (einfo,w)
 
 --------------------
 
@@ -111,7 +140,7 @@ systemN ∷ ∀ ε ζ ω μ .
            MakeProc ζ, OutputHandles ζ ω, HasCallStack,
            AsCreateProcError ε, AsFPathError ε, AsIOError ε, AsProcExitError ε,
            MonadError ε μ, HasCallStack) ⇒
-          CmdSpec → μ (ExitStatus, ω)
+          CmdSpec → μ (ExitInfo, ω)
 systemN c = devnull ≫ \ l → system l c
 
 --------------------
@@ -122,7 +151,7 @@ systemS ∷ ∀ ε ζ ω μ .
            MakeProc ζ, OutputHandles ζ ω, HasCallStack,
            AsCreateProcError ε, AsFPathError ε, AsIOError ε, AsProcExitError ε,
            MonadError ε μ, HasCallStack) ⇒
-          CmdSpec → μ (ExitStatus, ω)
+          CmdSpec → μ (ExitInfo, ω)
 systemS c = system stdin c
 
 ----------------------------------------
@@ -131,17 +160,25 @@ systemS c = system stdin c
      exit status is of a signal received. -}
 
 throwSig ∷ ∀ ε β η . (AsProcExitError ε, MonadError ε η) ⇒
-           CmdSpec → (ExitStatus, β) → η (Word8, β)
+           CmdSpec → Pid → (ExitInfo, β) → η (ExitInfo, β)
 
-throwSig _     (ExitVal x     , w) = return (x,w)
-throwSig cspec (ex@(ExitSig _), _) =
-  throwError $ asProcExitError cspec ex (𝕹,𝕹)
+{-
+throwSig cspec pd (view exitVal → ex@(ExitSig _),_)=
+  throwError $ asProcExitError cspec pd ex (𝕹,𝕹)
+throwSig _     _   (exstat@(view exitVal → ExitVal _), w) = return (exstat,w)
+-}
+
+throwSig cspec pd (exstat, w) =
+  let exVal = exstat ⊣ exitVal
+   in if exitWasSignalled exVal
+      then throwError $ asProcExitError cspec pd exVal (𝕹,𝕹)
+      else return (exstat,w)
 
 ----------
 
 throwSig' ∷ ∀ ε β η . (AsProcExitError ε, MonadError ε η) ⇒
-             CmdSpec → 𝔼 ε (ExitStatus, β) → η (Word8, β)
-throwSig' cspec = fromRight ∘ join ∘ fmap (throwSig cspec)
+            CmdSpec → Pid → 𝔼 ε (ExitInfo, β) → η (ExitInfo, β)
+throwSig' cspec pd = fromRight ∘ join ∘ fmap (throwSig cspec pd)
 
 ----------------------------------------
 
@@ -149,15 +186,16 @@ throwSig' cspec = fromRight ∘ join ∘ fmap (throwSig cspec)
      argument is always executed immediately after the process returns (whatever
      the exit value).
  -}
+
 doProc ∷ ∀ ε ζ ω σ μ .
          (MonadIO μ, MkInputStream σ,
           ToMaybeTexts ω, MakeProc ζ, OutputHandles ζ ω,
           AsProcExitError ε, AsCreateProcError ε, AsFPathError ε,
           AsIOError ε, Printable ε, MonadError ε μ) ⇒
-         μ () → σ → CmdSpec → μ (Word8, ω)
+         μ () → σ → CmdSpec → μ (ExitInfo, ω)
 doProc finally input cspec = do
-  result ← ѥ $ systemx input cspec
+  result ← systemx input cspec
   finally
-  throwSig' cspec result
+  throwSig' cspec (result ⊣ _1 ∘ pid) (𝕽 result)
 
 -- that's all, folks! ----------------------------------------------------------